# frontend/src/public/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "RIS API",
    "contact" : {
      "name" : "DigitalService GmbH des Bundes",
      "url" : "https://digitalservice.bund.de"
    },
    "version" : "1"
  },
  "servers" : [ {
    "url" : "https://testphase.rechtsinformationen.bund.de",
    "description" : "Public API (preview)"
  } ],
...
  "paths" : {
...
    "/v1/rechtsprechung/{documentNumber}" : {
      "get" : {
        "tags" : [ "Rechtsprechung" ],
        "summary" : "Judgment metadata (Metadaten eines Rechtsrechungsdokuments)",
        "description" : "The endpoint returns a single judgment (Rechtsrechungsdokument) from our database.",
        "operationId" : "getCaseLaw",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "STRE201770751"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RechtsprechungSchema"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/v1/rechtsprechung/{documentNumber}/{name}.{extension}" : {
      "get" : {
        "tags" : [ "Rechtsprechung" ],
        "summary" : "Caselaw resource",
        "description" : "Returns a specific resource of a particular caselaw.",
        "operationId" : "getImage",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "BDRE000800001"
        }, {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "image"
        }, {
          "name" : "extension",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "png", "jpg", "jpeg", "gif", "wmf", "emf", "bitmap" ]
          },
          "example" : "jpg"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/v1/rechtsprechung/{documentNumber}.zip" : {
      "get" : {
        "tags" : [ "Rechtsprechung" ],
        "summary" : "Decision ZIP (XML and attachments)",
        "description" : "Returns a case law decision, including attachments, as a ZIP archive.",
        "operationId" : "getCaseLawDocumentationUnitAsZip",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "^[a-zA-Z]{4}\\d{9}$"
          },
          "example" : "STRE201770751"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/zip" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StreamingResponseBody"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/v1/rechtsprechung/{documentNumber}.xml" : {
      "get" : {
        "tags" : [ "Rechtsprechung" ],
        "summary" : "Decision XML",
        "description" : "Returns a case law decision as XML. This content is used as a source for the HTML endpoint.",
        "operationId" : "getCaseLawDocumentationUnitAsXml",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "STRE201770751"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/xml" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/v1/rechtsprechung/{documentNumber}.html" : {
      "get" : {
        "tags" : [ "Rechtsprechung" ],
        "summary" : "Decision HTML",
        "description" : "Renders and returns a case law decision as HTML.",
        "operationId" : "getCaseLawDocumentationUnitAsHtml",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "STRE201770751"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/html" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/v1/rechtsprechung/changelog" : {
      "get" : {
        "tags" : [ "Rechtsprechung" ],
        "summary" : "Caselaw changelog",
        "description" : "Returns references of document changes that occurred in between two points in time.",
        "operationId" : "getChangelogs",
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChangelogResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/literature/changelog" : {
      "get" : {
        "tags" : [ "Literature" ],
        "summary" : "Literature changelog",
        "description" : "Returns references of document changes that occurred in between two points in time.",
        "operationId" : "getChangelogs_1",
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChangelogResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/legislation/changelog" : {
      "get" : {
        "tags" : [ "Legislation" ],
        "summary" : "Legislation changelog",
        "description" : "Returns references of document changes that occurred in between two points in time.",
        "operationId" : "getChangelogs_2",
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChangelogResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/document/lucene-search/case-law" : {
      "get" : {
        "tags" : [ "Case Law" ],
        "summary" : "Advanced search using Lucene query syntax",
        "operationId" : "caseLawSearch",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "description" : "The query filter based on Lucene query syntax",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, courtName, documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, courtName, documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|DATUM|courtName|documentNumber)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaCaseLawSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/v1/case-law" : {
      "get" : {
        "tags" : [ "Case Law" ],
        "summary" : "List and search decisions",
        "description" : "The endpoint returns a list of decisions from our database. The list is paginated and can be filtered and sorted.",
        "operationId" : "searchCaseLaw",
        "parameters" : [ {
          "name" : "fileNumber",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ecli",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "court",
          "in" : "query",
          "description" : "Filter by court name (Finanzgericht Münster, FG Münster, ArbG Köln) or court type (Finanzgericht, FG, ArbG). Supports both long and short names.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Filter by court name (Finanzgericht Münster, FG Münster, ArbG Köln) or court type (Finanzgericht, FG, ArbG). Supports both long and short names."
          }
        }, {
          "name" : "legalEffect",
          "in" : "query",
          "description" : "Corresponds to “Rechtskraft”, meaning that the decision referred to is legally binding.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Corresponds to “Rechtskraft”, meaning that the decision referred to is legally binding.",
            "enum" : [ "JA", "NEIN", "KEINE_ANGABE", "FALSCHE_ANGABE" ]
          }
        }, {
          "name" : "type",
          "in" : "query",
          "description" : "Filter by document type (Urteil, Versäumnisurteil, Entscheidung etc.). Multiple values may be specified as a comma-separated list or by repeating the parameter.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "description" : "Filter by document type (Urteil, Versäumnisurteil, Entscheidung etc.). Multiple values may be specified as a comma-separated list or by repeating the parameter.",
            "examples" : [ "Versäumnisurteil", "Entscheidung" ],
            "items" : {
              "type" : "string",
              "description" : "Filter by document type (Urteil, Versäumnisurteil, Entscheidung etc.). Multiple values may be specified as a comma-separated list or by repeating the parameter.",
              "examples" : [ "Versäumnisurteil", "Entscheidung" ]
            }
          }
        }, {
          "name" : "typeGroup",
          "in" : "query",
          "description" : "Extended filter by type group. Multiple values may be specified as a comma-separated list or by repeating the parameter.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "description" : "Extended filter by type group. Multiple values may be specified as a comma-separated list or by repeating the parameter.",
            "enum" : [ "Urteil", "Beschluss", "other" ],
            "items" : {
              "type" : "string",
              "description" : "Extended filter by type group. Multiple values may be specified as a comma-separated list or by repeating the parameter.",
              "enum" : [ "URTEIL", "BESCHLUSS", "OTHER", "Urteil", "Beschluss", "other" ]
            }
          }
        }, {
          "name" : "searchTerm",
          "in" : "query",
          "description" : "Searches for the given tokens in searchTerm. If searchTerm contains more than one token, all tokens must be in the document for the document to match.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Searches for the given tokens in searchTerm. If searchTerm contains more than one token, all tokens must be in the document for the document to match."
          }
        }, {
          "name" : "dateFrom",
          "in" : "query",
          "description" : "The from (greater than or equal) parameter returns all entities where date is later than, or equal to, the given date.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date",
            "description" : "The from (greater than or equal) parameter returns all entities where date is later than, or equal to, the given date."
          }
        }, {
          "name" : "dateTo",
          "in" : "query",
          "description" : "The to (less than or equal) parameter returns all entities where date is earlier than, or equal to, the given date.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date",
            "description" : "The to (less than or equal) parameter returns all entities where date is earlier than, or equal to, the given date."
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, courtName, documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, courtName, documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|DATUM|courtName|documentNumber)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaCaseLawSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/v1/case-law/{documentNumber}" : {
      "get" : {
        "tags" : [ "Case Law" ],
        "summary" : "Decision metadata",
        "description" : "The endpoint returns a single decision from our database.",
        "operationId" : "getCaseLaw_1",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "STRE201770751"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaseLawSchema"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/v1/case-law/{documentNumber}/{name}.{extension}" : {
      "get" : {
        "tags" : [ "Case Law" ],
        "summary" : "Caselaw resource",
        "description" : "Returns a specific resource of a particular caselaw.",
        "operationId" : "getImage_1",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "BDRE000800001"
        }, {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "image"
        }, {
          "name" : "extension",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "png", "jpg", "jpeg", "gif", "wmf", "emf", "bitmap" ]
          },
          "example" : "jpg"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/v1/case-law/{documentNumber}.xml" : {
      "get" : {
        "tags" : [ "Case Law" ],
        "summary" : "Decision XML",
        "description" : "Returns a case law decision as XML. This content is used as a source for the HTML endpoint.",
        "operationId" : "getCaseLawDocumentationUnitAsXml_1",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "STRE201770751"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/xml" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "StreamingResponseBody" : { },
...
      "ChangelogDeletedDocument" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string",
            "description" : "unique identifier of the document"
          },
          "@type" : {
            "type" : "string",
            "description" : "type of the document"
          }
        },
        "required" : [ "@id", "@type" ]
      },
      "ChangelogResponse" : {
        "type" : "object",
        "properties" : {
          "@type" : {
            "type" : "string"
          },
          "@context" : {
            "type" : "string"
          },
          "changed" : {
            "type" : "array",
            "description" : "Set of changed documents",
            "items" : {
              "$ref" : "#/components/schemas/ChangelogChangedDocument"
            },
            "uniqueItems" : true
          },
          "deleted" : {
            "type" : "array",
            "description" : "Set of deleted documents",
            "items" : {
              "$ref" : "#/components/schemas/ChangelogDeletedDocument"
            },
            "uniqueItems" : true
          },
          "allChanged" : {
            "type" : "boolean",
            "description" : "flag to communicate that the whole storage got rebuilt"
          }
        },
        "required" : [ "@context", "allChanged", "changed", "deleted" ]
      },
...
    }
  }
}